Removed DbSpellService
